requirements.txt
numpy
pandas
torch
transformers
wandb
Pillow
requests
matplotlib
torchmetrics